pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent_system"
version = "0.1.0"
description = "Recursive tool-calling agent with OpenAI LLM + HF SLM and KB logging"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "openai>=1.0.0",
  "requests>=2.31.0",
  "transformers>=4.51.0",
  "accelerate>=0.30.0",
  "torch",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_system*"]
